Direct import of all definitions in the Apple File System Reference
This is a complete import, verbatim of all C header definitions for the
Apple File System Reference as specified in the 2020-06-22 revision.
This does not currently compile as there are some dependency issues
with the ordering from the file.

This should be useful as a good way of verifying correctness when
ported to other languages.