Restructured B-Tree code to simplify generics and better handle non-leaf records
Leaf values now use an associated type to refer the appropriate key type
and records are now concrete structs for both the leaf and non-leaf
case. This greatly simplifies the generic parameters being used and
should make it much easier to finally complete support for the non-leaf
records.

The support for parsing APFS records has been disabled for the moment.