Clone B+ Tree records to avoid needing references and lifetimes
This is not ideal, but it the only immediately obvious solution to
returning a found B+ Tree record when recursively scanning across
multiple nodes. Ideally, this should be changed to just return a
reference to it to avoid extra copies of the data especially if they are
later modified in-place.