Implemented support for parsing the Space Manager Free Queue
This adds B+ Tree support for the Free Queue and allows it to be printed
out from the dump program. This does introduce a regression when
scanning some file systems as this B+ Tree utilizes ghost entries which
are not supported and trigger the dump program to panic when
encountered.