Implemented decoding and dump of EFI Jumpstart
This adds decoding of all structures in the EFI Jumpstart chapter. With
this, all structures in the Apple File System Specification should be
fully implemented for basic decode through the File-System Objects
chapter, however, only a visual verification through Object Maps is
completed. A few other later chapters are also known to be fully
implemented such as Reaper and Space Manager.