Added support for decoding the Object type values used in keybags
The object type value used for keybags uses all 32-bits overlapping what
is normally used for Storage type and flags, but due to the encrypted
nature of these objects, it's probably not used anyways. However, it
does make decoding them more complex and the current code is a bit of a
hack.